rebased on upstream
- use size_t/ssize_t in vfs
- bump vfs addons

+++ xbmc/addons/VFSEntry.h
@@ -44,8 +44,8 @@ namespace ADDON
     void* OpenForWrite(const CURL& url, bool bOverWrite);
     bool Exists(const CURL& url);
     int Stat(const CURL& url, struct __stat64* buffer);
-    unsigned int Read(void* ctx, void* lpBuf, int64_t uiBufSize);
-    int Write(void* ctx, void* lpBuf, int64_t uiBufSize);
+    ssize_t Read(void* ctx, void* lpBuf, size_t uiBufSize);
+    ssize_t Write(void* ctx, void* lpBuf, size_t uiBufSize);
     int64_t Seek(void* ctx, int64_t iFilePosition, int iWhence = SEEK_SET);
     int Truncate(void* ctx, int64_t size);
     void Close(void* ctx);
@@ -99,8 +99,8 @@ namespace ADDON
     virtual bool OpenForWrite(const CURL& url, bool bOverWrite);
     virtual bool Exists(const CURL& url);
     virtual int  Stat(const CURL& url, struct __stat64* buffer);
-    virtual unsigned int Read(void* lpBuf, int64_t uiBufSize);
-    virtual int Write(void* lpBuf, int64_t uiBufSize);
+    virtual ssize_t Read(void* lpBuf, size_t uiBufSize);
+    virtual ssize_t Write(void* lpBuf, size_t uiBufSize);
     virtual int64_t Seek(int64_t iFilePosition, int iWhence = SEEK_SET);
     virtual int Truncate(int64_t size);
     virtual void Close();
